@@ -1,16 +1,7 @@
-// This file is part of jsonapi
-//
-// jsonapi is distributed in two licenses: The Mozilla Public License,
-// v. 2.0 and the GNU Lesser Public License.
-//
-// jsonapi is distributed in the hope that it will be useful, but WITHOUT
-// ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
-// FOR A PARTICULAR PURPOSE.
-//
-// See LICENSE.txt for further information.
+// This Source Code Form is subject to the terms of the Mozilla Public
+// License, v. 2.0. If a copy of the MPL was not distributed with this
+// file, You can obtain one at https://mozilla.org/MPL/2.0/.
 
 module github.com/raohwork/jsonapi
 
-go 1.12
-
-require github.com/Ronmi/rtoolkit v0.0.4
+go 1.21
