@@ -1,13 +1,6 @@
-// This file is part of jsonapi
-//
-// jsonapi is distributed in two licenses: The Mozilla Public License,
-// v. 2.0 and the GNU Lesser Public License.
-//
-// jsonapi is distributed in the hope that it will be useful, but WITHOUT
-// ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
-// FOR A PARTICULAR PURPOSE.
-//
-// See LICENSE for further information.
+// This Source Code Form is subject to the terms of the Mozilla Public
+// License, v. 2.0. If a copy of the MPL was not distributed with this
+// file, You can obtain one at https://mozilla.org/MPL/2.0/.
 
 package apitool
 
@@ -16,19 +9,22 @@ import (
 	"context"
 	"encoding/json"
 	"io"
-	"io/ioutil"
 	"net/http"
 
 	"github.com/raohwork/jsonapi"
 )
 
 // EClient indicates something goes wrong at client side while calling remote jsonapi
+//
+// Deprecated: api client code is rewrited and placed in package callapi.
 var EClient = jsonapi.Error{Code: -1}.SetData("Client error")
 
 // Client is a helper to simplify the process of calling jsonapi
 //
 // Any error during the call process will immediately return as
 // jsonapi.InternalError.SetOrigin(the_error)
+//
+// Deprecated: api client code is rewrited and placed in package callapi.
 type Client interface {
 	// Synchronized call
 	//
@@ -73,6 +69,8 @@ type callResp struct {
 //
 // If any io error or json decoding error occurred, an
 // EClient.SetOrigin(the_error) returns.
+//
+// Deprecated: api client code is rewrited and placed in package callapi.
 func ParseResponse(resp *http.Response, result interface{}) error {
 	var res callResp
 	if err := json.NewDecoder(resp.Body).Decode(&res); err != nil {
@@ -95,6 +93,8 @@ func ParseResponse(resp *http.Response, result interface{}) error {
 // Call creates an Client to a jsonapi entry
 //
 // It will use http.DefaultClient if c == nil, but it's not recommended.
+//
+// Deprecated: api client code is rewrited and placed in package callapi.
 func Call(method, uri string, client *http.Client) Client {
 	c := client
 	if c == nil {
@@ -120,7 +120,7 @@ func Call(method, uri string, client *http.Client) Client {
 			return EClient.SetOrigin(err)
 		}
 		defer resp.Body.Close()
-		defer io.Copy(ioutil.Discard, resp.Body)
+		defer io.Copy(io.Discard, resp.Body)
 		defer cancel()
 
 		return ParseResponse(resp, result)
